pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "tamp_improv"
version = "0.1.0"
description = "TAMP physical improvisation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy==1.26.4",
    "pandas==2.2.3",
    "scipy==1.14.0",
    "matplotlib==3.9.2",
    "pandas-stubs==2.2.2.240909",
    "types-tabulate==0.9.0.20241207",
    "hydra-core==1.3.2",
    "omegaconf==2.3.0",
    "gymnasium==1.0.0",
    "stable_baselines3==2.4.0",
    "pybullet==3.2.7",
    "moviepy==1.0.3",
    "objaverse==0.1.7",
    "trimesh==4.7.1",
    "matplotlib-inline==0.1.7",
    "tqdm==4.66.5",
    "tomsutils@git+https://github.com/tomsilver/toms-utils.git",
    "tomsgeoms2d@git+https://github.com/tomsilver/toms-geoms-2d.git",
    "task-then-motion-planning@git+https://github.com/tomsilver/task-then-motion-planning.git",
    "relational-structs@git+https://github.com/tomsilver/relational-structs.git",
    "pybullet-blocks@git+https://github.com/tomsilver/pybullet-blocks.git",
    "pybullet_helpers@git+https://github.com/tomsilver/pybullet-helpers.git",
]

[project.optional-dependencies]
develop = [
    "black>=25.0.0",
    "docformatter==1.7.5",
    "isort>=5.13.2",
    "mypy==1.11.2",
    "pylint==3.3.1",
    "pytest==8.3.3",
    "pytest-pylint==0.21.0",
    "ipdb==0.13.13",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
tamp_improv = ["py.typed"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
py_version = 311
profile = "black"
multi_line_output = 2
skip_glob = ["venv/*"]
split_on_trailing_comma = true

[tool.mypy]
strict_equality = true
disallow_untyped_calls = true
warn_unreachable = true
exclude = ["venv/*", "scripts/*"]

[[tool.mypy.overrides]]
module = [
    "matplotlib.*",
    "networkx.*",
    "objaverse.*",
]
ignore_missing_imports = true
